refactor: separate handshake from init command
